+++ packages/s2-core/__tests__/spreadsheet/row-link-spec.ts
@@ -96,6 +96,7 @@ describe('Row Text Link Tests', () => {
 
     expect(linkFieldJump).toBeCalledWith({
       key: 'province',
+      cellData: rowNode,
       record: {
         province: '浙江',
         type: '笔',
@@ -124,6 +125,7 @@ describe('Row Text Link Tests', () => {
 
     expect(linkFieldJump).toBeCalledWith({
       key: 'city',
+      cellData: rowNode,
       record: {
         province: '浙江',
         city: '义乌1',
@@ -152,6 +154,7 @@ describe('Row Text Link Tests', () => {
 
     expect(linkFieldJump).toBeCalledWith({
       key: 'province',
+      cellData: rowNode,
       record: {
         province: '四川',
         type: '笔',