+++ packages/s2-core/src/common/interface/interaction.ts
@@ -135,7 +135,7 @@ export interface BrushSelectionInfo {
 
 export interface InteractionOptions {
   // record which row/col field need extra link info
-  linkFields?: string[];
+  linkFields?: string[] | ((meta: Node | ViewMeta) => boolean);
   // focus selected cell, like the spotlight
   selectedCellsSpotlight?: boolean;
   // highlight all row header cells and column header cells to which the hovered cell belongs