+++ packages/s2-core/src/utils/interaction/index.ts
@@ -1,5 +1,6 @@
 export * from './formatter';
 export * from './hover-event';
+export * from './link-field';
 export * from './merge-cell';
 export * from './select-event';
 export * from './state-controller';