+++ packages/s2-react/__tests__/data/strategy-data.ts
@@ -394,6 +394,7 @@ export const StrategyOptions: SheetComponentOptions = {
     return placeholder;
   },
   interaction: {
+    selectedCellsSpotlight: true,
     resize: {
       disable: (resizeInfo) => {
         return (
@@ -415,13 +416,15 @@ export const StrategyOptions: SheetComponentOptions = {
     text: [
       {
         mapping: (value, cellInfo) => {
-          const { meta } = cellInfo;
+          const { colIndex } = cellInfo;
           const isNilValue = isNil(value) || value === '';
-          if (meta?.fieldValue?.values[0][0] === value || isNilValue) {
+
+          if (colIndex === 0 || isNilValue) {
             return {
               fill: '#000',
             };
           }
+
           return {
             fill: isUpDataValue(value) ? '#FF4D4F' : '#29A294',
           };
@@ -432,6 +435,11 @@ export const StrategyOptions: SheetComponentOptions = {
       {
         position: 'right',
         mapping(value, cellInfo) {
+          const { colIndex } = cellInfo;
+          if (colIndex === 0) {
+            return null;
+          }
+
           return isUpDataValue(value)
             ? {
                 // icon 用于指定图标条件格式所使用的 icon 类型