+++ s2-site/examples/layout/custom/demo/custom-row-col-cell-width.ts
@@ -7,7 +7,7 @@ fetch(
   .then((dataCfg) => {
     const container = document.getElementById('container');
 
-    // 详情请查看: https://s2.antv.vision/zh/docs/manual/advanced/custom/cell-size
+    // 详情请查看: https://s2.antv.antgroup.com/zh/docs/manual/advanced/custom/cell-size
     const s2Options = {
       width: 600,
       height: 480,